Added existing files
Added files from legacy project. Current version features a still-to-be-developed Neural Network exectution method, a GUI to monitor the network being trained on, a training routine for training a network to predict the movement of a mouse and a Server/Client to collect and send mouse movement data for training.